requirements.txt
websockets>=12.0
aiohttp>=3.9
anthropic>=0.23
rich>=13.7
pandas>=2.1
numpy>=1.26
ta>=0.11
feedparser>=6.0
pytest>=8.0
pytest-asyncio>=0.23
pytest-mock>=3.12
aioresponses>=0.7
google-genai>=1.0
openai>=1.30
yfinance>=0.2
